Move AddressFormatter to subfolder
