+++ app/Event/EventGeoCoordinatesServiceProvider.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Event;
 
 use CultuurNet\UDB3\Address\CultureFeedAddressFactory;
-use CultuurNet\UDB3\Address\FullAddressFormatter;
-use CultuurNet\UDB3\Address\LocalityAddressFormatter;
+use CultuurNet\UDB3\Address\Formatter\FullAddressFormatter;
+use CultuurNet\UDB3\Address\Formatter\LocalityAddressFormatter;
 use CultuurNet\UDB3\Broadway\EventHandling\ReplayFilteringEventListener;
 use CultuurNet\UDB3\Container\AbstractServiceProvider;
 use CultuurNet\UDB3\Error\LoggerFactory;
