+++ tests/Organizer/CommandHandler/UpdateGeoCoordinatesCommandHandlerTest.php
@@ -7,16 +7,16 @@ namespace CultuurNet\UDB3\Organizer\CommandHandler;
 use Broadway\CommandHandling\Testing\CommandHandlerScenarioTestCase;
 use Broadway\EventHandling\EventBus;
 use Broadway\EventStore\EventStore;
+use CultuurNet\UDB3\Address\Address;
+use CultuurNet\UDB3\Address\Formatter\FullAddressFormatter;
+use CultuurNet\UDB3\Address\Formatter\LocalityAddressFormatter;
+use CultuurNet\UDB3\Address\Locality;
+use CultuurNet\UDB3\Address\PostalCode;
+use CultuurNet\UDB3\Address\Street;
 use CultuurNet\UDB3\Geocoding\Coordinate\Coordinates;
 use CultuurNet\UDB3\Geocoding\Coordinate\Latitude;
 use CultuurNet\UDB3\Geocoding\Coordinate\Longitude;
 use CultuurNet\UDB3\Geocoding\GeocodingService;
-use CultuurNet\UDB3\Address\Address;
-use CultuurNet\UDB3\Address\FullAddressFormatter;
-use CultuurNet\UDB3\Address\Locality;
-use CultuurNet\UDB3\Address\LocalityAddressFormatter;
-use CultuurNet\UDB3\Address\PostalCode;
-use CultuurNet\UDB3\Address\Street;
 use CultuurNet\UDB3\Model\ValueObject\Geography\CountryCode;
 use CultuurNet\UDB3\Organizer\Commands\UpdateGeoCoordinatesFromAddress;
 use CultuurNet\UDB3\Organizer\Events\GeoCoordinatesUpdated;
